LayoutTests/crypto/subtle/X25519-derive-bits-malformed-parameters-expected.txt

Test X25519 deriveBits operation with malformed parameters

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS crypto.subtle.deriveBits("X25519", privateKey, null) rejected promise  with TypeError: Member X25519Params.publicKey is required and must be an instance of CryptoKey.
PASS crypto.subtle.deriveBits({name: "X25519"}, privateKey, null) rejected promise  with TypeError: Member X25519Params.publicKey is required and must be an instance of CryptoKey.
PASS crypto.subtle.deriveBits({name: "X25519", public: true}, privateKey, null) rejected promise  with TypeError: Type error.
PASS crypto.subtle.deriveBits({name: "X25519", public: null}, privateKey, null) rejected promise  with TypeError: Type error.
PASS crypto.subtle.deriveBits({name: "X25519", public: undefined}, privateKey, null) rejected promise  with TypeError: Member X25519Params.publicKey is required and must be an instance of CryptoKey.
PASS crypto.subtle.deriveBits({name: "X25519", public: Symbol()}, privateKey, null) rejected promise  with TypeError: Type error.
PASS crypto.subtle.deriveBits({name: "X25519", public: { }}, privateKey, null) rejected promise  with TypeError: Type error.
PASS crypto.subtle.deriveBits({name: "X25519", public: 1}, privateKey, null) rejected promise  with TypeError: Type error.
PASS crypto.subtle.deriveBits({ name:"X25519", public:publicKey }, publicKey, null) rejected promise  with InvalidAccessError: CryptoKey doesn't support bits derivation.
PASS crypto.subtle.deriveBits({ name:"X25519", public:privateKey }, privateKey, null) rejected promise  with InvalidAccessError: The requested operation is not valid for the provided key.
PASS crypto.subtle.deriveBits({ name:"X25519", public:publicKey }, privateKey, 384) rejected promise  with OperationError: The operation failed for an operation-specific reason.
PASS successfullyParsed is true

TEST COMPLETE
